Lots of login related changes
First, fixed web.config because it was trying to create its own db
instead of using the one we created.  Then created a few users to test
with, then made new page TimeUpdate which is what you are directed to on
login.  Which will feed in current wait times for user to update.